7_service/fastapi_model_serving/requirements.txt
fastapi[standard]
uvicorn[standard]
ultralytics
opencv-python
Pillow
python-multipart